[Drawer component] starting with the persistent drawer option
This makes it a bit difficult, since the childs have fixed placed in the drawer component itself.
I'm not yet able to figure out how to do this without using a 'hacky' way like assuming that the last content will always be the column component where the content will be in.